[Decomp][Pet] Cohesion split: extract persistence to PetDatabase.cpp
